+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "formol",
-  "version": "2.5.11",
+  "version": "2.5.12-alpha.2",
   "description": "An opiniated react form framework.",
   "main": "lib/formol.js",
   "module": "src/index.js",
