+++ src/index.js
@@ -1,6 +1,7 @@
 // This is not optimal, we might need to find a better solution
 export { default } from './Formol'
 export { default as Field } from './Field'
+export * from './FieldBase'
 export { default as Inliner } from './Inliner'
 export { default as Conditional } from './Conditional'
 export { default as FormolContextWrapper, FormolContext } from './FormolContext'
@@ -21,3 +22,9 @@ export {
   diff,
   isModified,
 } from './utils/object'
+
+export { default as choicesAdapter } from './utils/choicesAdapter'
+export { default as memoizedChoices } from './utils/memoizedChoices'
+export { default as multipleAdapter } from './utils/multipleAdapter'
+
+console.log('FORMOL Exporting my own data NOW !!!!')
